Changes from duplicated lines
